+++ SurvivalGame/Assets/Items/Scripts/ItemData/ItemData.cs
@@ -15,6 +15,7 @@ public abstract class ItemData : ScriptableObject
     public string itemName;
     public ItemType itemType;
     public float weight;
+    public string itemDescription;
 
     public GameObject worldPrefab; // Prefab for when dropped into the world
     public Sprite uiIcon; // Icon for inventory UI